+++ scripts/templates/build_and_upload_batch.ps1
@@ -31,12 +31,13 @@ $vars = $winenv, $linuxenv
 foreach ($env in $vars) {
 
     Write-Host "Publishing for $($env.OSName)" -ForegroundColor DarkGreen
-    $source= Resolve-Path "..\..\src\SqlBuildManager.Console\bin\Debug\$frameworkTarget\$($env.BuildTarget)\publish"
+
     if($uploadonly -eq $false)
     {
         dotnet publish  "..\..\src\SqlBuildManager.Console\sbm.csproj" -r $env.BuildTarget --configuration Debug -f $frameworkTarget
     }
     
+    $source= Resolve-Path "..\..\src\SqlBuildManager.Console\bin\Debug\$frameworkTarget\$($env.BuildTarget)\publish"
     if($env.OSName -eq "Windows")
     {
         $version = (Get-Item "$($source)\sbm.exe").VersionInfo.ProductVersion  #Get version for Batch application