+++ src/AssemblyVersioning.cs
@@ -23,9 +23,8 @@
 //   1) create a new "Product Code" 
 //   2) Update the installer version to match the AssemblyVersion below.
 //   These can be found in SqlBuildManager.Setup -> Organize Your Setup -> General Information
-// ** Also, don't forget to update the change_notes.xml and .html files!
 
-[assembly: AssemblyVersion("14.0.0")]
-[assembly: AssemblyFileVersion("14.0.0")]
+[assembly: AssemblyVersion("14.0.1")]
+[assembly: AssemblyFileVersion("14.0.1")]